fixed #96, wrong format string for sprintf
git-svn-id: svn://svn.lighttpd.net/xcache/trunk@391 c26eb9a1-5813-0410-bd6c-c2e55f420ca7

+++ xc_malloc.c
@@ -120,7 +120,7 @@ static XC_MEM_INIT(xc_mem_malloc_init) /* {{{ */
 #define MINSIZE (ALIGN(sizeof(xc_mem_t)))
 	/* requires at least the header and 1 tail block */
 	if (size < MINSIZE) {
-		fprintf(stderr, "xc_mem_malloc_init requires %d bytes at least\n", MINSIZE);
+		fprintf(stderr, "xc_mem_malloc_init requires %lu bytes at least\n", (unsigned long) MINSIZE);
 		return NULL;
 	}
 	mem->shm = shm;
